+++ app/assets/javascripts/application.js
@@ -13,12 +13,13 @@
 //= require rails-ujs
 //= require turbolinks
 //= require jquery
+//= require select2
 //= require cocoon
 //= require semantic-ui
 //= require_tree .
 
-$(document).on('turbolinks:load',function() {
-    $('.message .close').on('click', function() {
+$(document).on('turbolinks:load', function () {
+    $('.message .close').on('click', function () {
         $(this).closest('.message').transition('fade');
     });
     $('.dropdown').dropdown();