+++ app/helpers/pullouts_helper.rb
@@ -1,0 +1,2 @@
+module PulloutsHelper
+end