+++ db/migrate/20180212080263_create_stores.rb
@@ -2,7 +2,6 @@ class CreateStores < ActiveRecord::Migration[5.1]
   def change
     create_table :stores do |t|
       t.string :name, null: false
-      t.references :brand, foreign_key: true
       t.text :address
       t.string :email
       t.string :contact_person, default: ""